requirements.txt
python-telegram-bot==20.6
web3==6.11.5
requests
